Fix IPC server shutdown (#429)
The problem was that the server might stay alive forever
due to the UnixSocket being open. What this change does is:

If there is no more requests (the requests channel closed),
then we just keep the server open if there are subscribers
or pending requests. Otherwise it is considered drained and
it closes itself to avoid leakage.

Co-authored-by: FelipeRosa <felipe.sgrosa@gmail.com>

Co-authored-by: FelipeRosa <felipe.sgrosa@gmail.com>