add http-rustls feature (#466)
The existing `http-tls` feature enables TLS support via
`hyper-tls` which in turn pulls in OpenSSL on Linux via
`native-tls`. OpenSSL is written in C and has a long
history of vulnerabilities caused by memory corruption.

The new `http-rustls` feature allows to choose a TLS
implementation that is written in Rust.